requirements.txt
django>=6.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels[daphne]>=4.0
channels-redis>=4.2
psycopg2-binary>=2.9
django-filter>=24.0
django-cors-headers>=4.3
drf-spectacular>=0.27
Pillow>=10.0
redis>=5.0
